Added authentication instructions and scripted authentication
I added this because the script example hangs at the bash invocation (that downloads attachments) when I do not have a valid bw session open in my shell. This might be self-evident for regular bw cli users, but not for me. I also thought it'd be nice to be able to use the BW_SESSION variable inside of the script by opening a new session at script invocation so you don't have to think about whether or not you have a session running.